2-4 skill star
apparently setting the skill star as 0x90 and putting it under "call kimitteSaiko"  causes the skill star to break so putting it as 0x150 and putting it under beat4 before the second kimitteSaiko works.
IDK why but it happens. also the async subs under the skill star is the reason it works apparently.